fix: move test to e2e module
