+++ e2e/Makefile
@@ -1,0 +1,8 @@
+
+.PHONY: test
+test: tools
+	go test ./...
+
+.PHONY: tools
+tools:
+	go mod download
