+++ e2e/doc.go
@@ -1,0 +1,2 @@
+// Package e2e contains end-to-end tests.
+package e2e
